FIX: Fixed fuzzy search for dates, id and emails.
- I wrongfully though that I needed to convert the data inside the function,
So I would need two schemas: one for strict find and other for fuzzy find.
- But the truth is that the input is already paarsed (and converted) by the schema,
if I specify that the input should comply to that schema.
- So (controller) getUsersHandler() function doesnt need to convert strings dates to Date objects since the schema already does that.
- So (service) findUsers() function doesnt need to convert strings dates to Date objects since the schema already does that.
(no need for parseValidDate() function since schema converts incomplete but valid dates to complete dates and I already support fuzzy fields like before, aftere, between).
- So no need of two separate schemas for findng users.

But was is really needed is to remove .uuid() and .eamil() from the schema fields;
since they enforce those types, and if I want to support fuzzy search, then they must be removed.

Note that for CREATING users Im being strict with the fields;
But for finding users I dont need to be strict since Im not alrering the database nor adding dirty inputs to it.