FEAT: Added read from file to seed database
Now, the scriot can also read from a file to get the users data it'll use to seed the database.
This allows same database seed for all the team, so testing is easier.
It supports json and csv as inoput files.
It also supports saving the successfully seeded users into a file (json or csv).
Note that with htis, is easier to translate json to csv and viceversa.
csv format can be opened in libreoffice to see it pretty and ordered.
It'll also easy new users creation, since data is tabulated;
so there's no need t repeat each fieldname as in json format.