FIX: Fixed script parsing
In csv files, it read the file correctly, but it sent the values TRIPLE QUOTED (e.g: """value""").
In json files, it failed to read the json file because of the comma (separator of json arrays)
Increased generator MAX_RETRIES from 20 to 40. To increase changes of getting valid data (accoding to constraints).