FEAT: Added execution summary and failed data save
Now, at the end of the execution, a summary is printed with the details (suceeded, failed)
If there is any failed attemp, then it can be also logged and saved into a special file.
This behavior is controlled by constants, so we can decide if we want to;
- Have a simple/full summary at the end of the execution.
- Save the failed data into a file.
The filename and location of the failed attemps folder/file is also controller via constants.
And is also customizable via script prompts, since it asks for format, filename and location.