FIX: Removed request validation logic in AUTH api
All zod schemas were removed and not "shape" is used.
Before, Zod was used to parse the request body, and zixed fields were used when destructuring the request. bode;
now AUTH service just forward the request to USERS for it to handle.
In this way, no field enforcing nor validation is made.
P.D: Still need to check if "shape" is not interfering neither