CHORE: Updated CODEOWNERS
- Put database seeder script as protected, so I can check if someone modify its code
(so I can also suggest implementation ways since I created it).
But allowed modifications to the default folder were generatede results will be
(So anyone can add its own files, and edit them too).
- Temporary owned Nginx config and AUTH service src folders.
It's because yesterday (250528) Filip and I haad issues with his PR that was changing (reverting)
files from AUTH and Nginx config, and it could be because I rewrote those branch git history.
So this is just a preventive way of checking that nothis get modifed unintentionally.