DOCS: Added most the words, acronyms and abbreviations we use (the correct ones)
Added to our CSpell dictiorary custom-word.txt most of the works we're using and which are flagged as "unknown words".
But I added only the "correct ones", the ones that are NOT correct (e.g: propernames NOT capitalized, variable names NOT cammelCased)
wont be added since they are wrong and should be corrected.