FEAT: Implemented call_site, stack_trace, service_name format strings
Now, logger has these enhacements:
- call_site:
Used by trace and debug mode for indicating function, file and line which triggered the message/logger.
It's useful to easily locate the origin, and it's structured to support ELK stack.
It's controlled by the constant SHOW_CALL_SITE (set to false to not print its info).
- stack_trace:
console.trace() includes the stack, useful for tracking bugs.
This is used by trace mode (always) and debug, error and fatal mode (only if LOG_STACK_TRACE=true).
It's controlled by the constant LOG_STACK_TRACE (set to false to not print its info).
- service name:
So each log has a field 'base' which indicated which service is logging. useful for find logs related to an specofic service.
It's controlled by the constant SERVICE_NAME (set it to empty to not print its info)
- format strings
As 'console.*()' do, it allows formatting a string by the use of placeholders (like printf).

The goal of logger is to replace the calls to 'console.*()' so we have a more structured logging system;
and later we could "easily" migrate to ELK stack (Elasticsearch, Logstash, Kibana) for log management.
This will be very useful to debug the code.