REFACT: main function in app.ts
Now main uses ft_fastifyErrorHandler for the error handler function;
instead of registering it as an arrow function and thefine it in place.
This makes the code cleaner, and all error stuff are in one place: error.ts

Also comment-in process.exit(1) because I DO WANT that the service fails in case of 'listen' fails.
Easier for detect in-use ports or something wrong with the system;
instead of having a service that ALWAYS try to start event when it can't.