FEAT: Implemented frienship logic and added helper function getUserOrThrow
The helper function is for following DRY methodology;
I was repeating a lot the logic for fetching a user from the database.
I also chaged somce "const user" to "let" so I can modify it later (and avoid creating two objects).
So I updated the existing services, and implemnted the new frienship ones.