FIX: Fixed issue with authGuard not finding authUser in users routes
The issue what that I used startWith(path) when checking for public routes
I did that because Swagger UI stuff start with /api/documentation/ + blahblah.css, etc
But that was also matching any user route, so the authUser was never attached to the request
because all user routes were not protected.
The eeror was geenrated in onlySelf.

The friend_request routes worked because it's a different endpoint/route.